refactor(mobile): partir citas_tab.dart en 3 modulos (fase 5b-2)
- citas_tab.dart    169  pantalla + _BarraTabs + _Badge
- lista_citas.dart  150  ListaCitas + _ListaAgrupada + _HeaderMes + _Vacio
- cita_card.dart    245  CitaCard + _BadgeEstatus + _BadgeSpec

Archivo original (552 lineas) movido a subcarpeta citas/.
Sin cambios funcionales. flutter analyze: 0 issues.
